+++ src/parse/liquidity.rs
@@ -1,11 +1,11 @@
-use std::f64::EPSILON;
+use std::f32::EPSILON;
 
 use crate::bitcoin::sats_to_btc;
 
 pub struct LiquidityClassification {
-    illiquid: f64,
-    liquid: f64,
-    highly_liquid: f64,
+    illiquid: f32,
+    liquid: f32,
+    highly_liquid: f32,
 }
 
 impl LiquidityClassification {
@@ -34,7 +34,7 @@ impl LiquidityClassification {
     }
 
     #[inline(always)]
-    pub fn split(&self, value: f64) -> LiquiditySplitResult {
+    pub fn split(&self, value: f32) -> LiquiditySplitResult {
         LiquiditySplitResult {
             all: value,
             illiquid: value * self.illiquid,
@@ -45,18 +45,18 @@ impl LiquidityClassification {
 
     /// Returns value in range 0.0..1.0
     #[inline(always)]
-    fn compute_illiquid(x: f64) -> f64 {
+    fn compute_illiquid(x: f32) -> f32 {
         Self::compute_ratio(x, 0.25)
     }
 
     /// Returns value in range 0.0..1.0
     #[inline(always)]
-    fn compute_liquid(x: f64) -> f64 {
+    fn compute_liquid(x: f32) -> f32 {
         Self::compute_ratio(x, 0.75)
     }
 
     #[inline(always)]
-    fn compute_ratio(x: f64, x0: f64) -> f64 {
+    fn compute_ratio(x: f32, x0: f32) -> f32 {
         let l = 1.0;
         let k = 25.0;
 
@@ -66,10 +66,10 @@ impl LiquidityClassification {
 
 #[derive(Debug, Default)]
 pub struct LiquiditySplitResult {
-    pub all: f64,
-    pub illiquid: f64,
-    pub liquid: f64,
-    pub highly_liquid: f64,
+    pub all: f32,
+    pub illiquid: f32,
+    pub liquid: f32,
+    pub highly_liquid: f32,
 }
 
 #[derive(Debug, Default)]
